Moved auth to local settings, translated last screen to Danish

+++ questionnaires/render/views.py
@@ -21,34 +21,36 @@ def home_refreshed(request):
 	return render_to_response('home.html', {}, context_instance=RequestContext(request))
 
 def home(request):
-	try:
-		sessions = Session.objects.filter(expire_date__gte=datetime.now())
-		for session in sessions:
-			data = session.get_decoded()
-			try: user = User.objects.filter(id=data.get('_auth_user_id', None))[0]
-			except: continue
-			if request.user == user:
-				session.delete()
-	except: pass
+	if settings.DO_AUTH:
+		try:
+			sessions = Session.objects.filter(expire_date__gte=datetime.now())
+			for session in sessions:
+				data = session.get_decoded()
+				try: user = User.objects.filter(id=data.get('_auth_user_id', None))[0]
+				except: continue
+				if request.user == user:
+					session.delete()
+		except: pass
 
-	if request.user.is_authenticated():
-		try: f = FirstLogin.objects.get(user=request.user)
-		except FirstLogin.DoesNotExist: f = FirstLogin.objects.create(user=request.user)
-		if f.firstLogin:
-			f.firstLogin = False
-			f.save()
-			return redirect('request_attributes')
-		identity.getAttributes(request.user, ['email'])
+		if request.user.is_authenticated():
+			try: f = FirstLogin.objects.get(user=request.user)
+			except FirstLogin.DoesNotExist: f = FirstLogin.objects.create(user=request.user)
+			if f.firstLogin:
+				f.firstLogin = False
+				f.save()
+				return redirect('request_attributes')
+			identity.getAttributes(request.user, ['email'])
 
 	return redirect(settings.ROOT_URL+'form/')
 
 
 @login_required
 def form(request):
-	auth = oauth2.getToken(request.user, 'connector_questionnaire.input_form_data')
-	if auth == None:
-		#show user site to authorize the form
-		return render_to_response('start_auth.html', {}, context_instance=RequestContext(request))
+	if settings.DO_AUTH:
+		auth = oauth2.getToken(request.user, 'connector_questionnaire.input_form_data')
+		if auth == None:
+			#show user site to authorize the form
+			return render_to_response('start_auth.html', {}, context_instance=RequestContext(request))
 	try:
 		Response.objects.get(user = request.user,form_version='1.0',variable_name='_submitted')
 		return HttpResponseRedirect(settings.ROOT_URL+'nochanges/');
